requirements.txt
boto3
python_dotenv
requests
bs4
autopep8
flask